+++ CHANGELOG.md
@@ -8,6 +8,7 @@ FEATURES
 * **New Check:** `S010`: check for `Schema` of only `Computed` enabled with `ValidateFunc` configured
 * **New Check:** `S011`: check for `Schema` of only `Computed` enabled with `DiffSuppressFunc` configured
 * **New Check:** `S012`: check for `Schema` that `Type` is configured
+* **New Check:** `S013`: check for `map[string]*Schema` that one of `Computed`, `Optional`, or `Required` is configured
 
 BUG FIXES